api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
pandas==2.2.0
openpyxl==3.1.2
python-multipart==0.0.9
python-dateutil==2.8.2
requests==2.31.0
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
